+++ app/src/main/java/com/bionic/kvt/serviceapp/models/Order.java
@@ -11,7 +11,7 @@ public class Order {
      * Order number [nummer in XML}
      * NotNull
      */
-    private int number;
+    private long number;
 
     private String orderType; //OrderType
     private Date date; //Datum
@@ -62,7 +62,7 @@ public class Order {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(long number) {
         this.number = number;
     }
